Update README, add requirements.txt; tidied up dataset generator script
