+++ requirements.txt
@@ -1,0 +1,6 @@
+pandas>=1.3
+numpy>=1.21
+faker>=13.0
+tqdm>=4.62
+python-dotenv>=0.19
+openai>=1.0
